[업그레이드] 파워, 시간 업그레이드 UI 작업 중

+++ Assets/02_script/data/table/Tables/SHTableServerUserUpgradeInfo.cs
@@ -10,8 +10,8 @@ using LitJson;
 public class SHTableServerUserUpgradeInfo : SHBaseTable
 {
     public string UserId;
-    public long MiningPowerLv;
-    public long ChargeTimeLv;
+    public int MiningPowerLv;
+    public int ChargeTimeLv;
 
     public SHTableServerUserUpgradeInfo()
     {
